apps/multi-modal/requirements.txt
streamlit>=1.40.0
openai>=1.40.0
python-dotenv>=1.0.0
